backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles==24.1.0",
    "aiosqlite==0.17.0",
    "alembic==1.15.1",
    "apscheduler==3.11.0",
    "asyncmy==0.2.9",
    "asyncpg==0.30.0",
    "bcrypt==4.0.1",
    "click==8.1.7",
    "croniter==6.0.0",
    "cryptography==45.0.2",
    "fastapi==0.115.2",
    "fastapi-limiter==0.1.6",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "httpx==0.27.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "langchain==1.2.0",
    "langchain-anthropic==1.3.0",
    "langchain-mcp-adapters==0.2.1",
    "langchain-openai==1.1.6",
    "loguru==0.7.3",
    "openpyxl==3.1.5",
    "pandas==2.2.2",
    "passlib==1.7.4",
    "pillow==11.0.0",
    "psutil==6.1.0",
    "psycopg2-binary==2.9.10",
    "pydantic-settings==2.5.2",
    "pydantic-validation-decorator==0.1.4",
    "pyjwt==2.9.0",
    "pymysql==1.1.2",
    "python-multipart==0.0.9",
    "redis==7.1.0",
    "rich==13.9.4",
    "sqlalchemy==2.0.45",
    "sqlglot[rs]==27.8.0",
    "typer==0.9.0",
    "user-agents==2.2.0",
    "uvicorn==0.30.6",
    "websockets==14.2",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
